Ticket #1548: SOPC_Variant_Dump: ByteString length invalid cast lead to inconsitent condition test
(size_t) variant->Value.Bstring.Length when Length is -1 leads to SIZE_MAX value and NULL dereference on Data field

+++ src/Common/opcua_types/sopc_builtintypes.c
@@ -4826,7 +4826,7 @@ SOPC_ReturnStatus SOPC_Variant_Dump(SOPC_Buffer* buf, const SOPC_Variant* varian
         {
             char hex2[3];
             // Display as simple hexadecimal raw data
-            for (size_t i = 0; i < (size_t) variant->Value.Bstring.Length; ++i)
+            for (int32_t i = 0; i < variant->Value.Bstring.Length; ++i)
             {
                 snprintf(hex2, 3, "%02" PRIX8, variant->Value.Bstring.Data[i]);
                 SOPC_Buffer_PrintCString(buf, hex2);
